+++ CHANGELOG.md
@@ -66,6 +66,7 @@
 + Added **Tooltip Tweaks Tooltip Format** - Show crop-specific farming fortune in tooltip. (contributed by appable)
 + Added command **/shcropspeedmeter** - Helps calculate the real farming fortune with the formula crops broken per block.
 + Added **Compost Low Notification** - Shows a notification as title when organic matter/fuel is low.
++ Added **Jacob's Contest Warning** - Show a warning shortly before a new jacob contest starts.
 
 ### Features from other Mods
 > *The following features are only there because I want them when testing SkyHanni features without other mods present.*